mksquashfs: fix progressbar for sort files
The progressbar was incorrectly enabled *after* writing files
if sort files were used.  This meant the progressbar would not
be updated whilst files were being written to the filesystem.

As this is the bulk of the work for Mksquashfs, this means
the progressbar basically didn't appear to update whilst
Mksquashfs was running, and then suddenly updated to 100%
at the end of Mksquashfs.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>